Inline route addition for cob spec
