+++ src/test/java/HttpServer/cobspec/routes/CobSpecUnitTests.java
@@ -1,5 +1,5 @@
 package HttpServer.cobspec.routes;
-import HttpServer.core.definer.IRouteDefiner;
+import HttpServer.core.definer.*;
 import HttpServer.core.router.Router;
 import HttpServer.core.socket.MockClient;
 import HttpServer.core.socket.MockTraffic;
@@ -12,6 +12,7 @@ import HttpServer.core.utility.QuietLogger;
 import org.junit.Before;
 import org.junit.Test;
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.function.Function;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -29,7 +30,45 @@ public class CobSpecUnitTests {
     @Before
     public void setup() {
         logger = new QuietLogger();
-        definer = new CobSpecRouteDefiner();
+        definer = new IRouteDefiner() {
+            private final Router router;
+
+            {
+                Router blankRouter = new Router();
+                FileRouteDefiner fileDefiner = new FileRouteDefiner("./cob_spec/public", blankRouter);
+                TeaRouteDefiner teaDefiner = new TeaRouteDefiner(fileDefiner.getRouter());
+                HashMap<String, String> redirectionMap = new HashMap<String, String>();
+                redirectionMap.put("/redirect", "/");
+                RedirectDefiner redirectDefiner = new RedirectDefiner(teaDefiner.getRouter(), redirectionMap);
+                this.router = redirectDefiner.getRouter();
+
+                // Options tests are a concern of the router, in my opinion. Not set with a definer.
+                Handler okHandler = new FunctionalHandler(200);
+                router.defineRoute("/method_options", "GET", okHandler);
+                router.defineRoute("/method_options", "HEAD", okHandler);
+                router.defineRoute("/method_options", "POST", okHandler);
+                router.defineRoute("/method_options", "OPTIONS", okHandler);
+                router.defineRoute("/method_options", "PUT", okHandler);
+                router.defineRoute("/method_options2", "GET", okHandler);
+                router.defineRoute("/method_options2", "OPTIONS", okHandler);
+                serveRoot();
+                serveForm();
+            }
+
+            @Override
+            public Router getRouter() {
+                return router;
+            }
+
+            private void serveForm() {
+                router.defineRoute("/form", "PUT", new FunctionalHandler(200));
+            }
+
+            private void serveRoot() {
+                router.defineRoute("/", "GET", new FunctionalHandler(200));
+                router.defineRoute("/", "HEAD", new FunctionalHandler(200));
+            }
+        };
         router = definer.getRouter();
         runner = (requestLine) -> {
             MockTraffic simpleGet = new MockTraffic().request(new String[]{
